backend/E_commerce_BC/requirements.txt
asgiref==3.6.0
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
sqlparse==0.4.3
